Tokenizer
A ".core" file is passed into the program and a stream of tokens are created (if valid). The program halts when finding an invalid token and reports the location of the invalid token to the console. If the token stream is successfully made, then the the resulting tokens are outputted to the console in integer format (the tokens in string form are mapped to integers). This was made in Java.